prototypes/corp-assistant/pyproject.toml
[project]
name = "seo-marketing"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.24.0",
    "aiosqlite>=0.22.1",
    "alembic>=1.18.0",
    "aspose-words>=25.12.0",
    "asyncpg>=0.31.0",
    "chromadb>=1.4.1",
    "content-types>=0.3.0",
    "fastapi[all]>=0.128.0",
    "faststream[msgspec,redis]>=0.6.5",
    "langchain>=1.2.3",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "langgraph-checkpoint-redis>=0.3.2",
    "markdown>=3.10",
    "markitdown[all]>=0.1.4",
    "msgspec>=0.20.0",
    "pydub>=0.25.1",
    "python-magic-bin>=0.4.14",
    "pytz>=2025.2",
    "sentence-transformers>=5.2.0",
    "sqlalchemy>=2.0.45",
    "tinytag>=2.2.0",
]

[tool.ruff]
line-length = 99
preview = true
exclude = [
    "migrations",
    "migration",
    "src/migrations",
    "alembic.ini",
    "**/__pycache__",
    "**/.mypy_cache",
    "**/.ruff_cache",
    "**/.venv"
]

[tool.ruff.lint]
select = [
    "F",
    "E",
    "W",
    "C90",
    "I",
    "N",
    "UP",
    "YTT",
    "ASYNC",
    "S",
    "BLE",
    "B",
    "A",
    "COM",
    "C4",
    "DTZ",
    "T10",
    "EXE",
    "ISC",
    "ICN",
    "G",
    "INP",
    "PIE",
    "T20",
    "PYI",
    "PT",
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SLOT",
    "SIM",
    "TID",
    "INT",
    "ARG",
    "TD",
    "FIX",
    "ERA",
    "PD",
    "PGH",
    "PL",
    "TRY",
    "FLY",
    "NPY",
    "AIR",
    "PERF",
    "FURB",
    "LOG",
    "RUF",
]
ignore = [
    "S105",
    "S101",
    "B008",
    "TRY201",
    "PLW3201",
    "N805",
    "PTH",
    "COM812",
    "ISC001",
    "TRY003",
    "PT013",
    "TID252",
    "RUF001",
    "RUF002",
    "RUF003",
    "ASYNC109",
]

[tool.ruff.lint.isort]
section-order = [
    "future",
    "typings",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]

[tool.ruff.lint.isort.sections]
"typings" = ['typing']

[tool.ruff.lint.pylint]
max-args = 15
max-returns = 10
max-branches = 30

# -- MyPy --
[tool.mypy]
ignore_missing_imports = true
exclude = [
    "migrations",
    "src/migration",
    "src/migrations",
    "__samples__",
    ".venv"
]

# -- Vulture --
[tool.vulture]
exclude = [
    ".vscode/",
    "__samples__/",
    "tests",
    ".venv"
]
ignore_names = [
    "post",
    "put",
    "delete",
    "head"
]
